# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CRM API Documentation",
        "description": "CRM API Documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/owner/project-timesheet/list": {
            "post": {
                "tags": [
                    "Project Timesheet"
                ],
                "summary": "List Project Timesheet",
                "description": "List Project Timesheet",
                "operationId": "ProjectTimesheetList",
                "parameters": [
                    {
                        "name": "start",
                        "in": "query",
                        "description": "no of record you already get",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "0"
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "no of record you want to get",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "10"
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "search by name, mobile, email",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": ""
                    },
                    {
                        "name": "project_id",
                        "in": "query",
                        "description": "Project ID",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": ""
                    },
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "User ID",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": ""
                    },
                    {
                        "name": "sort_by",
                        "in": "query",
                        "description": "Sort by column(eg. id, name, mobile, email)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": ""
                    },
                    {
                        "name": "sort_order",
                        "in": "query",
                        "description": "Sort order (asc or desc)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": ""
                    }
                ],
                "responses": {
                    "200": {
                        "description": "json schema",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Invalid Request"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/owner/project-timesheet/store": {
            "post": {
                "tags": [
                    "Project Timesheet"
                ],
                "summary": "Store a Project Timesheet",
                "description": "Project Timesheet",
                "operationId": "ProjectTimesheetStore",
                "parameters": [
                    {
                        "name": "project_id",
                        "in": "query",
                        "description": "Enter Project ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "1"
                    },
                    {
                        "name": "start_time",
                        "in": "query",
                        "description": "Enter Start Time",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "100"
                    },
                    {
                        "name": "end_time",
                        "in": "query",
                        "description": "Enter End Time",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "8"
                    },
                    {
                        "name": "notes",
                        "in": "query",
                        "description": "Enter Notes",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Add notes here"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "json schema",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Invalid Request"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/owner/project-timesheet/update": {
            "post": {
                "tags": [
                    "Project Timesheet"
                ],
                "summary": "Update a Project Timesheet",
                "description": "Project Timesheet",
                "operationId": "ProjectTimesheetUpdate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Enter Project ID",
                        "required": true,
                        "schema": {
                            "type": "number"
                        },
                        "example": "1"
                    },
                    {
                        "name": "project_id",
                        "in": "query",
                        "description": "Enter Project ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "1"
                    },
                    {
                        "name": "start_time",
                        "in": "query",
                        "description": "Enter Start Time",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "100"
                    },
                    {
                        "name": "end_time",
                        "in": "query",
                        "description": "Enter End Time",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "8"
                    },
                    {
                        "name": "notes",
                        "in": "query",
                        "description": "Enter Notes",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Add notes here"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "json schema",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Invalid Request"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/owner/project-timesheet/view": {
            "post": {
                "tags": [
                    "Project Timesheet"
                ],
                "summary": "View a Project Timesheet",
                "description": "Project Timesheet",
                "operationId": "ProjectTimesheetView",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Enter Project Timesheet ID",
                        "required": true,
                        "schema": {
                            "type": "number"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "json schema",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Invalid Request"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/owner/project-timesheet/change-status": {
            "post": {
                "tags": [
                    "Project Timesheet"
                ],
                "summary": "Change Project Timesheet Status",
                "description": "Project Timesheet",
                "operationId": "ProjectTimesheetChangeStatus",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Enter Project Timesheet ID",
                        "required": true,
                        "schema": {
                            "type": "number"
                        },
                        "example": "1"
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Enter Status(InPlanning, Running, Stopped, Completed)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "Active",
                                "Inactive"
                            ]
                        },
                        "example": "Active"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "json schema",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Invalid Request"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/owner/project-timesheet/destroy": {
            "post": {
                "tags": [
                    "Project Timesheet"
                ],
                "summary": "Destroy a Project Timesheet",
                "description": "Project Timesheet",
                "operationId": "ProjectTimesheetDestroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Enter Project Timesheet ID",
                        "required": true,
                        "schema": {
                            "type": "number"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "json schema",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Invalid Request"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}